improvements for load filters options
