+++ app/views/annuaire/search/_form.html.haml
@@ -1,6 +1,6 @@
 = form_with url: annuaire_search_path, local: true, class: 'fr-mb-3w' do |f|
   = f.hidden_field :institution_slug, value: params[:institution_slug].presence || (institution.slug if defined?(institution) && institution.present?)
-  .fr-grid-row.fr-grid-row--gutters{ data: { controller: 'stats', url: load_filter_options_team_index_path } }
+  .fr-grid-row.fr-grid-row--gutters{ data: { controller: 'stats', url: annuaire_load_filter_options_path } }
     .fr-col
       .fr-input-group.fr-pr-2v
         .autocomplete-field{ 'data-autocomplete-url-value': "#{annuaire_autocomplete_path}", 'data-controller': 'autocomplete', 'data-autocomplete-min-length': 3 }
